requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
flake8==5.0.4
idna==3.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.3
stripe==5.0.0
urllib3==1.26.12
